moved MobileMetricCard to components/
