+++ src/components/MobileMetricCard/MobileMetricCard.module.scss
@@ -1,5 +1,5 @@
-@import '~@santiment-network/ui/mixins.scss';
-@import '~@santiment-network/ui/variables.scss';
+@import '../../../node_modules/@santiment-network/ui/mixins';
+@import '../../../node_modules/@santiment-network/ui/variables';
 
 .metric {
   display: flex;
